+++ public/index.html
@@ -22,6 +22,7 @@
 </head>
 
 <body>
+  <%= htmlWebpackPlugin.options.embed %>
   <noscript>
     <strong>We're sorry but page doesn't work properly without JavaScript enabled. Please enable it to
       continue.</strong>
