+++ src/style/embed.scss
@@ -1,0 +1,1 @@
+@import "./main.scss";
