+++ src/style/index.scss
@@ -1,0 +1,9 @@
+@import './main.scss';
+
+// styles for main page (index.js)
+
+// ce spreminjas visino navbara
+body {
+  margin-top: 76px;
+  // background-color: #f5f5f0;
+}
